alikson/teleport/old_version/requirements.txt
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
demoji==0.4.0
emoji-country-flag==1.2.4
idna==2.10
Pyrogram==0.17.1
PySocks==1.7.1
pytz==2021.1
requests==2.25.1
TgCrypto==1.2.2
urllib3==1.26.5
xlrd==2.0.1
xlwt==1.3.0
Twisted~=21.2.0